test(kueue): split the restart harness and prove non-vacuity on the live cluster (fbiy-B2b)
